// lib/gen/models/GetTagResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MVC
 * MVC Assignment
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface GetTagResponse
 */
export interface GetTagResponse {
  /**
   *
   * @type {number}
   * @memberof GetTagResponse
   */
  id?: number;
  /**
   *
   * @type {string}
   * @memberof GetTagResponse
   */
  name?: string;
}

/**
 * Check if a given object implements the GetTagResponse interface.
 */
export function instanceOfGetTagResponse(value: object): value is GetTagResponse {
  return true;
}

export function GetTagResponseFromJSON(json: any): GetTagResponse {
  return GetTagResponseFromJSONTyped(json, false);
}

export function GetTagResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetTagResponse {
  if (json == null) {
    return json;
  }
  return {
    id: json['id'] == null ? undefined : json['id'],
    name: json['name'] == null ? undefined : json['name']
  };
}

export function GetTagResponseToJSON(json: any): GetTagResponse {
  return GetTagResponseToJSONTyped(json, false);
}

export function GetTagResponseToJSONTyped(value?: GetTagResponse | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }

  return {
    id: value['id'],
    name: value['name']
  };
}